[REFACTOR] #202 MyProfileView MemberProfileView 통합
