+++ Whistle/UI/Screens/Profile/MemberFollowListView.swift
@@ -167,9 +167,14 @@ extension MemberFollowListView {
     ScrollView {
       ForEach(filteredFollower, id: \.userName) { follower in
         NavigationLink {
-          MemberProfileView(userId: follower.followerId)
+          MyProfileView(
+            profileType: .member,
+            isFirstProfileLoaded: .constant(true),
+            userId: follower.followerId)
             .environmentObject(apiViewModel)
-            .id(UUID())
+//          MemberProfileView(userId: follower.followerId)
+//            .environmentObject(apiViewModel)
+//            .id(UUID())
         } label: {
           personRow(
             isFollowed: Binding(get: {
@@ -194,9 +199,14 @@ extension MemberFollowListView {
     ScrollView {
       ForEach(filteredFollowing, id: \.userName) { following in
         NavigationLink {
-          MemberProfileView(userId: following.followingId)
+          MyProfileView(
+            profileType: .member,
+            isFirstProfileLoaded: .constant(true),
+            userId: following.followingId)
             .environmentObject(apiViewModel)
-            .id(UUID())
+//          MemberProfileView(userId: following.followingId)
+//            .environmentObject(apiViewModel)
+//            .id(UUID())
         } label: {
           personRow(
             isFollowed: Binding(get: {
