+++ Whistle/UI/Screens/Profile/MyProfile/BookMarkedFeedView.swift
@@ -107,7 +107,10 @@ struct BookMarkedFeedView: View {
     }
     .navigationDestination(isPresented: $feedMoreModel.isRootStacked) {
       if !apiViewModel.bookmark.isEmpty {
-        MemberProfileView(userId: apiViewModel.bookmark[feedPlayersViewModel.currentVideoIndex].userId)
+        MyProfileView(
+          profileType: .member,
+          isFirstProfileLoaded: .constant(true),
+          userId: apiViewModel.bookmark[feedPlayersViewModel.currentVideoIndex].userId)
       }
     }
     .fullScreenCover(isPresented: $feedMoreModel.showReport, onDismiss: {
